+++ hydrated.json.d.ts
@@ -1,0 +1,3 @@
+import { HydratedEntry } from './source/types'
+declare const hydratedList: HydratedEntry[]
+export default hydratedList