+++ raw.json.d.ts
@@ -1,0 +1,3 @@
+import { RawEntry } from './source/types'
+declare const rawList: RawEntry[]
+export default rawList